+++ lib/generators/repres/dosser/platform/templates/app/controllers/portals_controller.rb
@@ -1,0 +1,16 @@
+class PlatformModuleName::Dosser::VersionModuleName::PortalsController < PlatformModuleName::Dosser::VersionModuleName::PresentationController
+
+  def show
+    render_ok collection: [
+      {
+        name: 'PlatformModuleName API version_name',
+        links: [
+          { rel: 'self',      href: platform_name_dosser_version_name_portal_url(Unidom::Common::SELF, format: request.format.symbol) },
+          { rel: 'canonial',  href: platform_name_dosser_version_name_portal_url(Unidom::Common::SELF, format: :json                ) },
+          { rel: 'alternate', href: platform_name_dosser_version_name_portal_url(Unidom::Common::SELF, format: :xml                 ) }
+        ]
+      }
+    ]
+  end
+
+end