+++ lib/generators/repres/dosser/platform/templates/app/controllers/presentation_controller.rb
@@ -1,0 +1,17 @@
+require 'unidom'
+
+class PlatformModuleName::Dosser::VersionModuleName::PresentationController < ActionController::Base
+
+  clear_helpers
+
+  include Repres::Dosser::Concerns::ResourcePresentation
+
+  layout nil
+
+  before_action :load_unidom
+
+  def load_unidom
+    initialize_unidom
+  end
+
+end