+++ lib/generators/repres/dosser/platform/templates/config/routes.rb
@@ -1,0 +1,3 @@
+  scope '/platform_name-api/version_name', module: 'platform_name/dosser/version_name', as: 'platform_name_dosser_version_name' do
+    resources :portals, only: :show
+  end